requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
charset-normalizer==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
Pillow==9.4.0
pyTelegramBotAPI==4.10.0
python-telegram-bot==13.15
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wincertstore==0.2
